+++ routes/v0/index.js
@@ -1,7 +1,7 @@
 const { Router } = require('express');
 const contacts = require('./contacts');
 const cases = require('./cases');
-const postSurveys = require('./postSurveys');
+const postSurveys = require('./post-surveys');
 
 const router = Router();
